% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canada_modis_data.R
\docType{data}
\name{canada_modis_data}
\alias{canada_modis_data}
\title{MODIS NDVI dataset}
\format{
A data frame with 33189 rows and 7 variables
}
\source{
Processed in 'data-process/03-ndvi-data-process.R'
}
\usage{
data(canada_modis_data)
}
\description{
A dataset containing MODIS NDVI data
}
\details{
\itemize{
  \item ID. Replicate plot name
  \item site. Chronosequence site
  \item sample. Letter referring to plot name
  \item Date. Date sample taken
}
}
\keyword{datasets}
