% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_solution.R
\name{compute_solution}
\alias{compute_solution}
\title{Compute one step of a linear differential equation}
\usage{
compute_solution(f, p, y_old, int_p_old, dt)
}
\arguments{
\item{f}{forcing function}

\item{p}{time-dependent coefficient of linear DE}

\item{y_old}{previous time value}

\item{int_p_old}{previous value of the accumulated integral of p(t)}

\item{dt}{timestep}
}
\value{
Vector of the new_y value and the new value of the accumulated integral p(t)
}
\description{
\code{compute_solution} solve one step of the differential equation y' + p(t)y = f(t)
}
