% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffusivity.R
\docType{data}
\name{diffusivity}
\alias{diffusivity}
\title{Computed site diffusivity}
\format{
A data frame with 967 rows and 7 variables
}
\source{
Processed in 'data-process/05-process-diffusivity.R'
}
\usage{
data(diffusivity)
}
\description{
A dataset containing MODIS NDVI data
}
\details{
\itemize{
  \item ID. Replicate plot name
  \item layer. Depth of sample (10 or 5 cm)
  \item Date. Date sample taken
  \item rate. first derivative (dC/dt)
  \item sdv. second derivative (d^2C/dz^2)
  \item diffusivity. computed diffusivity
  \item real. Gap filled or not?
}
}
\keyword{datasets}
