% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euler_solution.R
\name{euler_solution}
\alias{euler_solution}
\title{Solve one step of the euler differential equation y' = f(y) (can be a system of equations)}
\usage{
euler_solution(y_old, rates, dt)
}
\arguments{
\item{y_old}{previous time value}

\item{rates}{the right hand side}

\item{dt}{timestep}
}
\value{
Vector of the new_y value and the new value of the accumulated integral p(t)
}
\description{
\code{euler_solution} solve one step of the differential equation y' = f(y)
}
