% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_data.R
\docType{data}
\name{field_data}
\alias{field_data}
\title{Recorded field values from Canada chronosequence sites.}
\format{
A nested data frame.  Data is processed in 06-field-data-prepare.R
}
\usage{
data(field_data)
}
\description{
\itemize{
  \item Year. Represents year since the site was burned
  \item field_data. Collected field measurements
}
}
\keyword{datasets}
