% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formBMatrix.R
\name{formBMatrix}
\alias{formBMatrix}
\title{Internal function to create the first difference matrix}
\usage{
formBMatrix(newDim, kernelSize)
}
\arguments{
\item{newDim}{Number of rows in the matrix}

\item{kernelSize}{Number of BRDF kernels}
}
\value{
The B matrix of first order difference. See Quaife and Lewis 2010
}
\description{
\code{formBmatrix} Computes the B matrix (currently the first order difference) used in the BRDF kernel inversion.  Accounts for multiple observations in a given day.
}
\examples{

B <- formBmatrix(365,3)

}
