% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpp_timeseries.R
\docType{data}
\name{gpp_timeseries}
\alias{gpp_timeseries}
\title{Recorded GPP values from MODIS}
\format{
A data frame with 74508 observations.  See the file appears_data.csv for information about the specific latitude and longitude for each site.  Data is processed in 04-gpp-timeseries-process.R
}
\usage{
data(gpp_timeseries)
}
\description{
\itemize{
  \item ID. Specific code name for the site and plot
  \item site. Represents year since the site was burned
  \item sample. Which particular plot this belongs to
  \item Date. Date of GPP measurment
  \item gpp. GPP values (kg C m-2 day)
}
}
\keyword{datasets}
