% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsvd_compute.R
\name{gsvd_compute}
\alias{gsvd_compute}
\title{Compute the Generalized Singular Value Decomposition for two matrices}
\usage{
gsvd_compute(K, B)
}
\arguments{
\item{kernel_values}{A matrix of kernel weights from function kernel_matrix}
}
\value{
GSVD decomposed matrices
}
\description{
\code{gsvd_compute} Computes solution for a GSVD matrix decomposition
}
\examples{

# To be filled in later
}
