% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsvd_norm.R
\name{gsvd_norm}
\alias{gsvd_norm}
\title{Compute the solution for the kernel weights for a GSVD decomposition}
\usage{
gsvd_norm(gsvdResult, lambda_df, rho)
}
\arguments{
\item{gsvdResult}{GSVD decomposition}

\item{lambda_df}{smoothness parameter - can be a vector of values - must be a data frame}

\item{rho}{right hand side of the equation - is a data frame with the band, value, and measurement}
}
\value{
Data frame that contains the residual and solution norms with the associated lambda and reflectance band
}
\description{
\code{gsvd_norm} Computes residual and solution norm for a GSVD matrix decomposition
}
\examples{

# To be filled in later
}
