% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsvd_optimize.R
\name{gsvd_optimize}
\alias{gsvd_optimize}
\title{Compute the solution for the kernel weights for a GSVD decomposition}
\usage{
gsvd_optimize(gsvdResult, rho, sigma)
}
\arguments{
\item{gsvdResult}{GSVD decomposition}

\item{rho}{right hand side of the equation - is a data frame with the band, value, and measurement}

\item{sigma}{uncertainty that we want to target for the RMSE}
}
\value{
A vector of lambda values that optimize for each band.  Converged = True - able to determine the norm.  FALSE = max number of iterations reached.
}
\description{
\code{gsvd_optimize} Computes residual and solution norm for a GSVD matrix decomposition
}
\examples{

# To be filled in later
}
