% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_matrix_canada.R
\name{kernel_matrix_canada}
\alias{kernel_matrix_canada}
\title{Compute the matrix for BRDF inversion}
\usage{
kernel_matrix_canada(year, in_data)
}
\arguments{
\item{year}{Current year of data}

\item{in_data}{Input data to analyze}
}
\value{
A list that contains the K, B, and rho matrix, and time points
}
\description{
\code{kernel_matrix_canada} Computes the B, K, and rho matrices for a given site in preparation to do the GSVD
}
\examples{
# TBD
}
