% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lseq.R
\name{lseq}
\alias{lseq}
\title{Compute the a logarithmically spaced sequence}
\source{
From package emdbook: \url{https://www.rdocumentation.org/packages/emdbook/versions/1.3.10/topics/lseq}
}
\usage{
lseq(from = 1, to = 1e+05, length.out = 6)
}
\arguments{
\item{from}{Starting value}

\item{to}{Ending Value}

\item{length.out}{Size of the sequence}
}
\value{
Vector of logarithmic points
}
\description{
\code{lseq} Computes solution for a GSVD matrix decomposition
}
\examples{

lseq(1,100000,5)
}
