% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lst_data.R
\docType{data}
\name{lst_data}
\alias{lst_data}
\title{Recorded land surface temperature values from MODIS}
\format{
A data frame with 17496 observations.  See the file appears_data.csv for information about the specific latitude and longitude for each site.  Data is processed in 01-lst-data-process.R
}
\usage{
data(lst_data)
}
\description{
\itemize{
  \item ID. Specific code name for the site and plot
  \item site. Represents year since the site was burned
  \item sample. Which particular plot this belongs to
  \item Date. Date of LST measurment
  \item lst. LST values (K) - given on an 8 day timescale)
}
}
\keyword{datasets}
