% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lst_timeseries.R
\docType{data}
\name{lst_timeseries}
\alias{lst_timeseries}
\title{Smoothed land surface temperature values from MODIS}
\format{
A data frame with 71220 observations.  Data is processed in 01b-lst-timeseries-process.R
}
\usage{
data(lst_timeseries)
}
\description{
\itemize{
  \item ID. Specific code name for the site and plot
  \item site. Represents year since the site was burned
  \item sample. Which particular plot this belongs to
  \item Date. Date of LAST measurment
  \item lst. LST values (K)
}
}
\keyword{datasets}
