% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling_data.R
\docType{data}
\name{modeling_data}
\alias{modeling_data}
\title{Collected modeling datasets from Canada chronosequence sites.}
\format{
A nested data frame.  Data is processed in 07-prepare-data-model.R
}
\usage{
data(modeling_data)
}
\description{
\itemize{
  \item ID. Specific code name for the site and plot
  \item site. Represents year since the site was burned
  \item depth. Depth of measurement (5, 10, 30 cm)
  \item data. Data used for modeling, includes the site, sample, Date, gpp, swc, soilT
}
}
\keyword{datasets}
