% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling_param_data.R
\docType{data}
\name{modeling_param_data}
\alias{modeling_param_data}
\title{Collected modeling datasets and model parameters from Canada chronosequence sites.}
\format{
A nested data frame.  Data is processed in 08-parameters-model.R
}
\usage{
data(modeling_param_data)
}
\description{
\itemize{
  \item ID. Specific code name for the site and plot
  \item site. Represents year since the site was burned
  \item depth. Depth of measurement (5, 10, 30 cm)
  \item data. Data used for modeling, includes the site, sample, Date, gpp, swc, soilT
  \item parameters. Model parameters used at each site.
}
}
\keyword{datasets}
