% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndvi_timeseries.R
\docType{data}
\name{ndvi_timeseries}
\alias{ndvi_timeseries}
\title{Smoothed ndvi values from MODIS}
\format{
A data frame with 71220 observations.  Data is processed in 03e-ndvi-timeseries-process.R
}
\usage{
data(ndvi_timeseries)
}
\description{
\itemize{
  \item ID. Specific code name for the site and plot
  \item site. Represents year since the site was burned
  \item sample. Which particular plot this belongs to
  \item Date. Date of LAST measurment
  \item ndvi. ndvi_values
}
}
\keyword{datasets}
