% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlxb_jz.R
\name{nlxb_jz}
\alias{nlxb_jz}
\title{Parameter estimate using the Nash equivalent of LM algorithm.}
\source{
Function nlxb in the nlsr pacakge
}
\usage{
nlxb_jz(
  formula,
  start,
  trace = FALSE,
  data = NULL,
  lower = -Inf,
  upper = Inf,
  masked = NULL,
  weights = NULL,
  control = list()
)
}
\arguments{
\item{formula}{The formula (expression we use for data assimilation)}

\item{start}{Start values for parameters}

\item{trace}{Should we report progress}

\item{data}{What data will we use for assimilation?}

\item{lower}{lower bound for parameters}

\item{upper}{Upper bound for parameters}

\item{masked}{Parameters not used}

\item{weights}{weights on the data}

\item{control}{A list of controls for the algorithm.}
}
\description{
\code{nlxb_jz} Determines a parameter estimate.  I had to modify if from nlxb so it doesn't print out the progress messages.
}
\seealso{
nlxb from the nlsr pacakge
}
