% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_time_convert.R
\name{parameter_time_convert}
\alias{parameter_time_convert}
\title{Update a parameter value to a daily timescale}
\usage{
parameter_time_convert(parameters)
}
\arguments{
\item{name}{name of parameters}

\item{units}{timescale to convert - one of hr^-1 or yr^-1}

\item{value}{current value of parameter}
}
\value{
Data frame that contains the updated name and value
}
\description{
\code{parameter_time_convert} Updates a parameter value to a daily timestep (from hourly or yearly)
}
\examples{
 # TBD
}
