% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoother_compute.R
\name{smoother_compute}
\alias{smoother_compute}
\title{Report the GSVD solution for a given year associated with Land Surface temperature}
\usage{
smoother_compute(in_data, uncertainty_val)
}
\arguments{
\item{in_data}{input data for LST from MODIS for a given year. Has columns Date and column named measurement}

\item{uncertainty_val}{associated uncertainty with the given measurement - data frame with column sigma and the type of uncertainty:
- rmse: root mean square error
- rmse_percent: = rmse / mean (for signal to noise
- solution: solution norm
- residual: residual norm}
}
\value{
Data frame that contains the smoothed LST for a given day of the year
}
\description{
\code{smoother_compute} Computes computes the smoothed timeseries of GPP from a GSVD decomposition
}
\examples{

# To be filled in later
}
