% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil_temp_timeseries.R
\docType{data}
\name{soil_temp_timeseries}
\alias{soil_temp_timeseries}
\title{Modeled soil temperature values}
\format{
A nested data frame with 12 rows.  Data is processed in 05b-soil-ibutton-timeseries.R
}
\usage{
data(soil_temp_timeseries)
}
\description{
\itemize{
  \item ID. Specific code name for the site and plot
  \item layer. Depth of sample (5, 10, 30 cm)
  \item out_coeffs. Fit values from empirical fit
  \item T_soil. Fitted soil temperature at that depth
}
}
\keyword{datasets}
