% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil_temperature_iButton_data.R
\docType{data}
\name{soil_temperature_iButton_data}
\alias{soil_temperature_iButton_data}
\title{Measured soil temperature data from iButton measurements}
\format{
A nested data frame with 3177 observations and 5 variables.  Data is processed in 05-process-ibutton-data.R
}
\usage{
data(soil_temperature_iButton_data)
}
\description{
\itemize{
  \item ID. Specific code name for the site and plot
  \item Date. Date of measurements
  \item layer. Depth of sample (L5, L10, L30 cm)
  \item T_soil. Fitted soil temperature at that depth
  \item source. Instrument used (currently iButton only)
}
}
\keyword{datasets}
