% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swc_timeseries.R
\docType{data}
\name{swc_timeseries}
\alias{swc_timeseries}
\title{Smoothed soil water content values from MODIS}
\format{
A data frame with 71220 observations.  Data is processed in 04-swc-timeseries-process.R
}
\usage{
data(swc_timeseries)
}
\description{
\itemize{
  \item ID. Specific code name for the site and plot
  \item site. Represents year since the site was burned
  \item sample. Which particular plot this belongs to
  \item Date. Date of LAST measurment
  \item swc. swc values (%)
}
}
\keyword{datasets}
